data/requirements.txt
aiohttp>=3.8.0
supabase>=2.3.5
python-dotenv>=1.0.0